cvfeedback/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
Django==5.1.6
django-extensions==4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.9.0
multidict==6.4.4
openai==1.82.0
pillow==11.1.0
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
pydotplus==2.0.2
PyJWT==2.9.0
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
webencodings==0.5.1
yarl==1.20.0
